requirements.txt
# ===============================================
# FACE ATTENDANCE SYSTEM - COMPLETE REQUIREMENTS
# ===============================================

# Core Backend Framework
Django==5.1.4
djangorestframework==3.15.2
django-cors-headers==4.6.0
django-filter==24.2
django-extensions==3.2.3
django-environ==0.11.2
python-dotenv==1.0.1

# Database & ORM
psycopg2-binary==2.9.9
pgvector==0.2.5

# Caching & Sessions
redis==5.0.1
django-redis==5.4.0
hiredis==2.2.3

# Message Queue & Task Processing
pika==1.3.2
celery==5.4.0
django-celery-beat==2.7.0
django-celery-results==2.5.1
kombu==5.4.0

# Authentication & Security
PyJWT==2.9.0
djangorestframework-simplejwt==5.3.0
django-allauth==0.62.0
cryptography==43.0.1
argon2-cffi==23.1.0
django-otp==1.3.0

# API Documentation
drf-spectacular==0.27.2
drf-spectacular-sidecar==2024.11.1

# ===============================================
# AI/ML & Computer Vision Stack
# ===============================================

# Core ML Libraries
torch==2.5.1
torchvision==0.20.1
torchaudio==2.5.1
tensorflow==2.18.1

# Computer Vision
opencv-python==4.10.0.84
opencv-contrib-python==4.10.0.84
Pillow==10.4.0
scikit-image==0.24.0
imageio==2.36.0

# Face Detection & Recognition (simplified for initial setup)
mtcnn==0.1.1
deepface==0.0.79
facenet-pytorch==2.5.3

# Scientific Computing
numpy==1.26.4
scipy==1.13.1
pandas==2.2.2
scikit-learn==1.5.0
matplotlib==3.9.2
seaborn==0.13.2

# Model Optimization
onnx==1.16.1
onnxruntime==1.15.1
onnxruntime-gpu==1.18.1
tensorrt==8.6.1
openvino==2024.3.0

# ===============================================
# Web & API Technologies
# ===============================================

# Web Server & ASGI
gunicorn==23.0.0
uvicorn==0.32.1
whitenoise==6.8.2
daphne==4.0.0

# FastAPI for AI Services
fastapi==0.101.1
pydantic==2.1.1
aiofiles==23.1.0

# HTTP & Networking
requests==2.32.3
urllib3==2.2.2
httpx==0.28.1
websockets==14.1

# Data Serialization
marshmallow==3.23.2
jsonschema==4.23.0
PyYAML==6.0.2

# ===============================================
# Development & Testing
# ===============================================

# Testing Framework
pytest==8.3.4
pytest-django==4.9.0
pytest-cov==6.0.0
pytest-mock==3.14.0
factory-boy==3.3.1
faker==33.0.0

# Code Quality
black==24.10.0
isort==5.13.2
flake8==7.1.1
pylint==3.3.2
mypy==1.13.0
bandit==1.7.10

# Development Tools
ipython==8.28.0
jupyter==1.1.1
notebook==7.3.1
ipdb==0.13.13
watchdog==6.0.0

# ===============================================
# Monitoring & Logging
# ===============================================

# Metrics & Monitoring
prometheus-client==0.21.1
django-prometheus==2.3.1
psutil==6.1.0

# Logging
structlog==24.4.0
colorlog==6.9.0
sentry-sdk==2.19.0
loguru==0.7.0

# Health Checks
django-health-check==3.18.0

# ===============================================
# Storage & File Handling
# ===============================================

# Object Storage
boto3==1.35.80
botocore==1.35.80
minio==7.2.7

# File Processing
python-magic==0.4.27
python-slugify==8.0.1

# ===============================================
# Utilities & Helpers
# ===============================================

# Date & Time
python-dateutil==2.9.0
pytz==2024.2
arrow==1.3.0

# Validation & Parsing
validators==0.28.0
phonenumbers==8.13.55
email-validator==2.2.0

# System & OS
click==8.1.7
rich==13.9.4
tqdm==4.67.1

# ===============================================
# Camera & Streaming
# ===============================================

# Video Processing
ffmpeg-python==0.2.0
imageio-ffmpeg==0.5.1

# Real-time Communication
socketio
python-socketio

# ===============================================
# Edge Computing & Hardware
# ===============================================

# GPU Acceleration (NVIDIA)
pycuda==2023.1
cupy-cuda12x==13.6.0



# ===============================================
# Production Dependencies
# ===============================================

# Process Management
supervisor==4.2.5

# Performance
gevent==24.11.1
greenlet==3.1.1

# Memory Management
memory-profiler==0.61.0
objgraph==3.6.0

# ===============================================
# Security & Compliance
# ===============================================

# Encryption
pycryptodome==3.21.0

# GDPR & Privacy
django-anonymizer==0.5

# Rate Limiting
django-ratelimit==4.1.0

# ===============================================
# Mobile & API Integration
# ===============================================

# Push Notifications
pyfcm==1.5.4

# QR Code Generation
qrcode==7.4.2

# ===============================================
# Analytics & Reporting
# ===============================================

# Data Analysis
plotly==5.24.1
dash==2.18.2

# Report Generation
reportlab==4.2.5
openpyxl==3.1.5
xlsxwriter==3.2.0

# ===============================================
# Version Compatibility Notes
# ===============================================
# Python 3.11+ Required
# Django 5.1+ with latest security patches
# CUDA 12.0+ for GPU acceleration
# PostgreSQL 15+ with pgvector extension
# Redis 7.0+ for optimal performance
# Ubuntu 20.04+ or equivalent for production
# Updated to latest stable versions as of December 2024